Create README.md
Performed EDA on superstore dataset of more than 9000 rows and 19 columns. This project had three csv files of sales, product and customer data. Merged them into one single file for better analysis and understanding. Plotted graphs for the same to understand the relation between different aspects of the dataset. Found out the most profitable region and state and most and least items sold in by the store. There is also a conclusion section which provides a summary overview of the entire dataset from total profit to mean discount. The graphs are plots help in better understanding of the dataset